PM / devfreq: imx8m-ddrc: Measure bandwidth with perf
The imx8m ddrc has a performance monitoring block attached which can
be used to measure bandwidth usage and automatically adjust frequency.

There is already a perf driver for the ddr pmu so instead of
implementing a devfreq-events driver use the in-kernel perf API to
implement get_dev_status directly.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>