tcp: authopt: add crypto initialization
The crypto_shash API is used in order to compute packet signatures. The
API comes with several unfortunate limitations:

1) Packet signatures must be computed in softirq context
2) Allocating a crypto_shash can sleep and must be done in user context.
3) Packet signatures use dynamic "traffic keys" which require exclusive
access to crypto_shash for crypto_setkey.

The solution is to allocate one crypto_shash for each possible cpu for
each algorithm at setsockopt time. The per-cpu tfms are then borrowed
from softirq context, signatures are computed tand the tfms are
returned.

This is similar to what is done for tcp md5.

Signed-off-by: Leonard Crestez <cdleonard@gmail.com>