tcp: authopt: Initial support for TCP_AUTHOPT_FLAG_ACTIVE
This can be used to determine if tcp authentication option is actually active
on the current connection.

TCP Authentication can be enabled but inactive on a socket if keys are only
configured for destinations other than the peer.

A listen socket with authentication enabled will return other sockets with
authentication enabled on accept(). If no key is configured for the peer of an
accepted socket then authentication will be inactive.

Signed-off-by: Leonard Crestez <cdleonard@gmail.com>